Add support for uploading files and reorganize protocols
